Add initial support for view-only API data for navigation purposes - #2607
