+++ src/components/ApiPageReleaseNotes.tsx
@@ -24,7 +24,11 @@ export class ApiPageReleaseNotes extends React.Component<RouteComponentProps<IAp
     if (apis.length > 1) {
       const apiCards = apis.map((apiDesc: IApiDescription) => {
         const { name, shortDescription, urlFragment, vaInternalOnly } = apiDesc;
-        const dashUrlFragment = urlFragment.replace('_', '-')
+        const dashUrlFragment = urlFragment.replace('_', '-');
+
+        if (apiDesc.isViewOnly) {
+          return null;
+        }
 
         return (
           <Flag key={name} name={`hosted_apis.${urlFragment}`}>
